Moved plot_forecast function to wrapper
